Added permissions functionaility

+++ backend/src/resolvers/Mutation.js
@@ -3,7 +3,7 @@ const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 
-const { setCookie } = require('../utils');
+const { setCookie, hasPermission } = require('../utils');
 const { makeANiceEmail, transport } = require('../mail');
 
 const Mutations = {
@@ -35,9 +35,18 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id, title }`);
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`);
+
     // 2. check if they own that item, or have the permission
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission),
+    );
+
+    if (!ownsItem || !hasPermissions) {
+      throw new Error(`You don't have permission to do that`);
+    }
+
     // 3. delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -165,6 +174,38 @@ const Mutations = {
 
     // return the new user
     return updatedUser;
+  },
+  async updatePermissions(parent, args, ctx, info) {
+    // check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in!');
+    }
+
+    // query the current user
+    const currentUser = await ctx.db.query.user(
+      {
+        where: {
+          id: ctx.request.userId,
+        },
+      },
+      info,
+    );
+
+    // check if they have permissions to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE']);
+
+    // update the permissions
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions,
+          },
+        },
+        where: { id: args.userId },
+      },
+      info,
+    );
   },
 };
